+++ jest.config.js
@@ -2,9 +2,7 @@
 
 const common = {
   modulePathIgnorePatterns: [
-    '<rootDir>/packages/outline/npm',
-    '<rootDir>/packages/outline-react/npm',
-    '<rootDir>/packages/outline-extensions/npm',
+    '/npm',
   ],
 };
 
@@ -13,20 +11,12 @@ module.exports = {
     {
       ...common,
       displayName: 'unit',
-      testPathIgnorePatterns: [
-        '<rootDir>/packages/outline-playground/',
-        '<rootDir>/packages/outline-react/src/__tests__/utils',
-      ],
+      testMatch: ['**/__tests__/unit/**/*.js'],
     },
     {
       ...common,
       displayName: 'e2e',
-      testPathIgnorePatterns: [
-        '<rootDir>/packages/outline/',
-        '<rootDir>/packages/outline-react/',
-        '<rootDir>/packages/outline-extensions/',
-        '<rootDir>/packages/outline-playground/__tests__/utils/',
-      ],
+      testMatch: ['**/__tests__/e2e/**/*.js'],
     },
   ],
 };
