+++ packages/outline-playground/__tests__/e2e/Placeholder-test.js
@@ -11,7 +11,7 @@ import {
   assertHTMLSnapshot,
   assertSelection,
   focusEditor,
-} from './utils';
+} from '../utils';
 
 describe('Placeholder', () => {
   initializeE2E({chromium: true, webkit: true, firefox: true}, (e2e) => {
