+++ packages/outline-playground/__tests__/e2e/TextFormatting-test.js
@@ -14,7 +14,7 @@ import {
   keyUpCtrlOrMeta,
   repeat,
   focusEditor,
-} from './utils';
+} from '../utils';
 
 describe('TextFormatting', () => {
   initializeE2E({chromium: true, webkit: true, firefox: true}, (e2e) => {
