+++ packages/outline-react/src/__tests__/unit/OutlineSelection-test.js
@@ -42,7 +42,7 @@ import {
   undo,
   redo,
   pastePlain,
-} from './utils';
+} from 'outline-react/src/__tests__/utils';
 
 describe('OutlineSelection tests', () => {
   let container = null;
